packages/agent/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "idp_code_agent" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/packages/agent/html'
 --cov-report xml:'../../coverage/packages/agent/coverage.xml'
 --html='../../reports/packages/agent/unittests/html/index.html'
 --junitxml='../../reports/packages/agent/unittests/junit.xml'
"""

[tool.hatch.build.targets.wheel]
packages = [ "idp_code_agent" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 120
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[project]
name = "idp_code.agent"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
 "bedrock-agentcore>=1.6.3",
 "strands-agents>=1.36.0",
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"
